kafka integration
implemented consumers and producers for gateway-response-topic and payment-topic for both core and integration service

+++ core/src/main/java/com/tandapayinterview/core/kafka/PaymentRequest.java
@@ -1,8 +1,14 @@
 package com.tandapayinterview.core.kafka;
 
+/**
+ * payload for payment request
+ * @param paymentId
+ * @param payableAmount
+ * @param mssdn
+ */
 public record PaymentRequest(
         String paymentId,
-        Double payableAmount,
+        Float payableAmount,
         String mssdn
 ) {
 }
